+++ instrumentation/spring/spring-web/spring-web-3.1/library/src/test/java/io/opentelemetry/instrumentation/spring/web/v3_1/SpringWebHeaderSelectorTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.spring.webmvc.v6_0;
+package io.opentelemetry.instrumentation.spring.web.v3_1;
 
 import static io.opentelemetry.api.common.AttributeKey.stringArrayKey;
 import static java.util.Arrays.asList;
@@ -14,28 +14,34 @@ import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.instrumentation.api.config.IncludeExclude;
 import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension;
 import io.opentelemetry.instrumentation.testing.junit.LibraryInstrumentationExtension;
-import jakarta.servlet.Filter;
-import jakarta.servlet.http.HttpServletResponse;
+import java.io.ByteArrayInputStream;
+import java.io.InputStream;
+import java.net.URI;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
-import org.springframework.mock.web.MockHttpServletRequest;
-import org.springframework.mock.web.MockHttpServletResponse;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.client.ClientHttpRequest;
+import org.springframework.http.client.ClientHttpRequestInterceptor;
+import org.springframework.http.client.ClientHttpResponse;
+import org.springframework.http.client.SimpleClientHttpRequestFactory;
 
-class SpringWebMvcHeaderCaptureTest {
+class SpringWebHeaderSelectorTest {
 
   @RegisterExtension
   static final InstrumentationExtension testing = LibraryInstrumentationExtension.create();
 
   @Test
   void capturesHeadersMatchingSelectorPatterns() throws Exception {
-    Filter filter =
-        SpringWebMvcTelemetry.builder(testing.getOpenTelemetry())
+    ClientHttpRequestInterceptor interceptor =
+        SpringWebTelemetry.builder(testing.getOpenTelemetry())
             .setRequestHeaders(IncludeExclude.builder().setIncluded("X-Test-*").build())
             .setResponseHeaders(IncludeExclude.builder().setExcluded("x-secret-*").build())
             .build()
-            .createServletFilter();
+            .createInterceptor();
 
-    handleRequest(filter);
+    sendRequest(interceptor);
 
     Attributes attributes = testing.waitForTraces(1).get(0).get(0).getAttributes();
     assertThat(attributes.get(stringArrayKey("http.request.header.x-test-request")))
@@ -49,14 +55,14 @@ class SpringWebMvcHeaderCaptureTest {
   @Test
   @SuppressWarnings("deprecation") // testing deprecated API
   void capturesHeadersConfiguredByName() throws Exception {
-    Filter filter =
-        SpringWebMvcTelemetry.builder(testing.getOpenTelemetry())
+    ClientHttpRequestInterceptor interceptor =
+        SpringWebTelemetry.builder(testing.getOpenTelemetry())
             .setCapturedRequestHeaders(asList("X-Test-Request", "Authorization"))
             .setCapturedResponseHeaders(singletonList("X-Test-Response"))
             .build()
-            .createServletFilter();
+            .createInterceptor();
 
-    handleRequest(filter);
+    sendRequest(interceptor);
 
     Attributes attributes = testing.waitForTraces(1).get(0).get(0).getAttributes();
     assertThat(attributes.get(stringArrayKey("http.request.header.x-test-request")))
@@ -74,14 +80,14 @@ class SpringWebMvcHeaderCaptureTest {
   @Test
   @SuppressWarnings("deprecation") // testing deprecated API
   void deprecatedSettersMatchHeaderNamesLiterally() throws Exception {
-    Filter filter =
-        SpringWebMvcTelemetry.builder(testing.getOpenTelemetry())
+    ClientHttpRequestInterceptor interceptor =
+        SpringWebTelemetry.builder(testing.getOpenTelemetry())
             .setCapturedRequestHeaders(singletonList("*"))
             .setCapturedResponseHeaders(singletonList("*"))
             .build()
-            .createServletFilter();
+            .createInterceptor();
 
-    handleRequest(filter);
+    sendRequest(interceptor);
 
     Attributes attributes = testing.waitForTraces(1).get(0).get(0).getAttributes();
     // "*" is dropped while the selector is built, so it captures nothing; Authorization is in the
@@ -92,21 +98,48 @@ class SpringWebMvcHeaderCaptureTest {
         .noneMatch(key -> key.getKey().startsWith("http.response.header."));
   }
 
-  private static void handleRequest(Filter filter) throws Exception {
-    MockHttpServletRequest request = new MockHttpServletRequest("GET", "/test");
-    request.addHeader("X-Test-Request", "request-value");
-    request.addHeader("X-Secret-Token", "secret-value");
-    request.addHeader("Authorization", "secret-value");
+  private static void sendRequest(ClientHttpRequestInterceptor interceptor) throws Exception {
+    // a real request implementation, since the HttpRequest interface has gained methods over the
+    // supported spring web version range
+    ClientHttpRequest request =
+        new SimpleClientHttpRequestFactory()
+            .createRequest(URI.create("http://localhost:8080/test"), HttpMethod.GET);
+    request.getHeaders().add("X-Test-Request", "request-value");
+    request.getHeaders().add("X-Secret-Token", "secret-value");
+    request.getHeaders().add("Authorization", "secret-value");
 
-    MockHttpServletResponse response = new MockHttpServletResponse();
+    TestClientHttpResponse response = new TestClientHttpResponse();
+    response.getHeaders().add("X-Test-Response", "response-value");
+    response.getHeaders().add("X-Secret-Token", "secret-value");
 
-    filter.doFilter(
-        request,
-        response,
-        (req, resp) -> {
-          HttpServletResponse httpResponse = (HttpServletResponse) resp;
-          httpResponse.addHeader("X-Test-Response", "response-value");
-          httpResponse.addHeader("X-Secret-Token", "secret-value");
-        });
+    interceptor.intercept(request, new byte[0], (req, body) -> response);
+  }
+
+  private static class TestClientHttpResponse implements ClientHttpResponse {
+
+    private final HttpHeaders headers = new HttpHeaders();
+
+    @Override
+    public HttpStatus getStatusCode() {
+      return HttpStatus.OK;
+    }
+
+    @Override
+    public String getStatusText() {
+      return "OK";
+    }
+
+    @Override
+    public void close() {}
+
+    @Override
+    public InputStream getBody() {
+      return new ByteArrayInputStream(new byte[0]);
+    }
+
+    @Override
+    public HttpHeaders getHeaders() {
+      return headers;
+    }
   }
 }
